requirements.txt
# Core Streamlit and web framework
streamlit==1.33.0

# Data manipulation and analysis
pandas==2.2.2
numpy==1.26.4

# Data visualization
plotly==5.22.0

# Database and ORM
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# Supabase integration
supabase==2.7.4

# Environment variable management
python-dotenv==1.0.1

# Authentication and security
cryptography==42.0.5

# Additional utilities
python-dateutil==2.9.0.post0

# HTTP client pins (compat with supabase 2.7.4 and chromadb 0.5.3)
httpx==0.27.2
httpcore==1.0.5

langchain==0.3.27
langchain-community==0.3.27
langchain-google-genai==2.1.9
langchain-text-splitters==0.3.9
langsmith==0.4.5
chromadb==0.5.3
pypdf==4.2.0
beautifulsoup4==4.12.3



